Make rspec make_assertion compatible with new code
The make_assertion call in FlexMock::Expectation provides a second
argument, @location, to the make_assertion call. As a result this is no
longer compatible with the FlexMock RSpec integration. This leads to a
test failure in the rspec integration tests.

It is not clear to me if the backtrace can and should be passed to
rspec, so I have opted to only provide a placeholder argument for it.