do not unnecessarily mock #class and #kind_of? in flexmock_based_on
While it is necessary to do it for 'full' mocks, partial mocks
(which are one of the biggest users of #flexmock_based_on through
the usage of partials_are_based) seldom need to do it.

This solves many spurious 'redefinition' warnings when running
under -w, when mocking new instances of a class and partials_are_based
is set, e.g.

   flexmock(MyClass).new_instances....

would (unnecessarily) redefine #class and #kind_of?, but since MyClass
is a permanent object, it would do it over and over again, which leads
to said redefinition warnings.