in partial mocks, find the original method through metaprogramming
Use a combination of Method#owner and Method#super_method.

It takes care of corner cases, such as adding a new method
in the call chain, between the time the mock was created and
the time the method was called.

It also removes the need to store the "right" method on mock
definition. Finding the right method is now done by looking
at properties of the call chain instead.